pyproject.toml
[tool.poetry]
name = "backend-api"
version = "0.1.0"
description = ""
authors = ["SALHI Mohamed <salhi.mohamed.sd@gmail.com>"]
readme = "README.md"
packages = [{include = "core"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.28.0"
django-environ = "^0.12.0"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.4.0"
django-filter = "^24.3"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.0"
black = "^25.1.0"
pylint = "^3.3.4"
mypy = "^1.15.0"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
factory-boy = "^3.3.3"
django-test-plus = "^2.2.4"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
